Initial Structure of Dia
The idea is to instantiate dia.core.DiaCore providing an implementation of dia.interfaces.AbstractDescriptiveRepository.
After that, provide at least one implementation of dia.predictive.interfaces.AbstractPredictiveSystem through method add_predictive_system of DiaCore.

From here, there is a lot of work to do. We will start working for dia.predictive.systems.statistical and to detail with more effort the interface of the predictive systems and the format of the recommendations returned for each system.